+++ src/types.ts
@@ -164,3 +164,8 @@ export interface MarginParams {
   optionDynamicPercentageShortMaintenance: number;
   optionShortPutCapPercentage: number;
 }
+
+// Only support margin accounts for now.
+export enum ProgramAccountType {
+  MarginAccount = "MarginAccount",
+}